parameter type is now int (was long long int)
- a routine has a computational and memory complexity. These are defined
  in terms of the parameters of the routine. In general, the parameters
  which have influcence are scalars, usually describing data sizes or
  number of iterations.
- In both these roles, the plain int datatype is sufficiently large. So,
  long long int has never made sense, it is now replaced by int

git-svn-id: https://hcl.ucd.ie/repos/pmm/trunk@285 9ba544ec-ad43-0410-a9a8-8e97fe1879fe

+++ src/pmm_argparser.h
@@ -39,7 +39,7 @@ typedef struct pmm_view_options {
     int plot_params_index;
 
     int *slice_i_arr;
-    long long int *slice_val_arr;
+    int *slice_val_arr;
     int slice_arr_size;
 
     char *model_file;
